WIP: Receipts refactoring
