--- crates/svm-codec/src/receipt/testing/receipt/mod.rs
@@ -1,8 +1,0 @@
-mod deploy_receipt;
-mod exec_app;
-mod helpers;
-mod spawn_app;
-
-pub use deploy_receipt::{decode_template_receipt, ClientTemplateReceipt};
-pub use exec_app::{decode_exec_receipt, ClientExecReceipt};
-pub use spawn_app::{decode_app_receipt, ClientAppReceipt};
